requirements.txt
python-dotenv
cryptography==41.0.2
requests
